Finished implementing on/off feature, big re-write
I finally finished implementing the on/off feature. In order to
accomplish this I needed a LOT of help from the user ZyX in
StackOverflow and ZyX_I in BitBucket.

I really appreciate all of this help, couldn't have done it without him.
In order to be able to accomplish this I needed to do pretty much a
total rewrite, mostly of function names, variable names, structure etc.

The magic happened here:
stackoverflow.com/questions/13751883
chat.stackoverflow.com/rooms/20762/discussion-between-eduan-and-zyx

Once again, many thanks to ZyX for all his help, now and in the past.